Allow dragging and dropping due tasks onto the calendar
